# lecture_3/rest/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Swagger Example"
  },
  "host": "localhost",
  "basePath": "/api/",
  "tags": [
    {
      "name": "user"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/v1/users": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Создание нового пользователя",
        "operationId": "createUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "запрос на создание пользователя",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUserRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User successfully created",
            "schema": {
              "$ref": "#/definitions/CreateUserResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "409": {
            "description": "User already exists",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          }
        }
      }
    },
    "/v1/users/{id}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Создание нового пользователя",
        "operationId": "getUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "id пользователя",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully created",
            "schema": {
              "$ref": "#/definitions/GetUserResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Обновление пользователя",
        "operationId": "updateUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "id пользователя",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "запрос на обновление пользователя",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully updated",
            "schema": {
              "$ref": "#/definitions/UpdateUserResponse"
            }
          },
          "204": {
            "description": "No content",
            "schema": {
              "type": "object"
            }
          },
          "403": {
            "description": "Forrbidden",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Удаление пользователя",
        "operationId": "deleteUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "id пользователя",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully deleted",
            "schema": {
              "$ref": "#/definitions/GetUserResponse"
            }
          },
          "204": {
            "description": "No content",
            "schema": {
              "type": "object"
            }
          },
          "403": {
            "description": "Forrbidden",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "ErrorMessage": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "CreateUserRequest": {
      "type": "object",
      "required": [
        "name",
        "email",
        "password"
      ],
      "properties": {
        "name": {
          "type": "string",
          "example": "user123"
        },
        "email": {
          "type": "string",
          "example": "user123@mail.ru"
        },
        "password": {
          "type": "string"
        },
        "avatar_photo_url": {
          "type": "string"
        }
      }
    },
    "CreateUserResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        }
      }
    },
    "GetUserResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "user123"
        },
        "email": {
          "type": "string",
          "example": "user123@mail.ru"
        },
        "avatar_photo_url": {
          "type": "string"
        }
      }
    },
    "UpdateUserRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "user123"
        },
        "email": {
          "type": "string",
          "example": "user123@mail.ru"
        },
        "password": {
          "type": "string"
        },
        "avatar_photo_url": {
          "type": "string"
        }
      }
    },
    "UpdateUserResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "user123"
        },
        "email": {
          "type": "string",
          "example": "user123@mail.ru"
        },
        "avatar_photo_url": {
          "type": "string"
        }
      }
    }
  }
}