# lecture_3/rest/server/models/create_user_response.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// CreateUserResponse create user response
//
// swagger:model CreateUserResponse
type CreateUserResponse struct {

	// id
	ID int64 `json:"id,omitempty"`
}

// Validate validates this create user response
func (m *CreateUserResponse) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this create user response based on context it is used
func (m *CreateUserResponse) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *CreateUserResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *CreateUserResponse) UnmarshalBinary(b []byte) error {
	var res CreateUserResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}