# lecture_3/rest/server/models/error_message.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// ErrorMessage error message
//
// swagger:model ErrorMessage
type ErrorMessage struct {

	// message
	Message string `json:"message,omitempty"`
}

// Validate validates this error message
func (m *ErrorMessage) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this error message based on context it is used
func (m *ErrorMessage) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *ErrorMessage) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *ErrorMessage) UnmarshalBinary(b []byte) error {
	var res ErrorMessage
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}