# lecture_3/rest/server/models/get_user_response.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// GetUserResponse get user response
//
// swagger:model GetUserResponse
type GetUserResponse struct {

	// avatar photo url
	AvatarPhotoURL string `json:"avatar_photo_url,omitempty"`

	// email
	// Example: user123@mail.ru
	Email string `json:"email,omitempty"`

	// id
	// Example: 1
	ID int64 `json:"id,omitempty"`

	// name
	// Example: user123
	Name string `json:"name,omitempty"`
}

// Validate validates this get user response
func (m *GetUserResponse) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this get user response based on context it is used
func (m *GetUserResponse) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *GetUserResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *GetUserResponse) UnmarshalBinary(b []byte) error {
	var res GetUserResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}