# lecture_3/rest/server/models/update_user_response.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// UpdateUserResponse update user response
//
// swagger:model UpdateUserResponse
type UpdateUserResponse struct {

	// avatar photo url
	AvatarPhotoURL string `json:"avatar_photo_url,omitempty"`

	// email
	// Example: user123@mail.ru
	Email string `json:"email,omitempty"`

	// id
	// Example: 1
	ID int64 `json:"id,omitempty"`

	// name
	// Example: user123
	Name string `json:"name,omitempty"`
}

// Validate validates this update user response
func (m *UpdateUserResponse) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this update user response based on context it is used
func (m *UpdateUserResponse) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *UpdateUserResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *UpdateUserResponse) UnmarshalBinary(b []byte) error {
	var res UpdateUserResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}